feat(sdk): extract api.ts client logic into standalone soroban-guard-sdk
