+++ packages/sdk/src/client.ts
@@ -1,0 +1,120 @@
+declare const process: any;
+import type { ScanResult, StellarNetwork, ScanResponse, ScanQuota } from './types'
+
+export class ApiError extends Error {
+  public retryAfter?: number
+
+  constructor(
+    public status: number,
+    message: string,
+    retryAfter?: number,
+  ) {
+    super(message)
+    this.name = 'ApiError'
+    this.retryAfter = retryAfter
+  }
+}
+
+export class TimeoutError extends Error {
+  constructor() {
+    super('The scan timed out after 30 seconds. Please try again.')
+    this.name = 'TimeoutError'
+  }
+}
+
+export interface SorobanGuardClientOptions {
+  baseUrl?: string
+  apiKey?: string
+}
+
+export class SorobanGuardClient {
+  private baseUrl: string
+  private apiKey?: string
+
+  constructor(options: SorobanGuardClientOptions = {}) {
+    const defaultUrl = typeof process !== 'undefined' ? (process.env.NEXT_PUBLIC_API_URL ?? 'http://localhost:3001') : 'http://localhost:3001'
+    this.baseUrl = (options.baseUrl ?? defaultUrl).replace(/\/$/, '')
+    this.apiKey = options.apiKey ?? (typeof process !== 'undefined' ? process.env.API_SECRET_KEY : undefined)
+  }
+
+  /**
+   * Submit source code to the Soroban Guard API for scanning.
+   * @param source - Contract source code or identifier
+   * @param network - Optional Stellar network to target
+   * @returns Scan result including findings and optional quota info
+   * @throws {ApiError} On HTTP errors or rate limiting
+   */
+  public async scan(source: string, network?: StellarNetwork): Promise<ScanResult> {
+    const body = { source }
+
+    const headers: Record<string, string> = { 'Content-Type': 'application/json' }
+    if (network) headers['X-Network'] = network.name
+    if (this.apiKey) {
+      headers['Authorization'] = `Bearer ${this.apiKey}`
+    }
+
+    const MAX_ATTEMPTS = 3
+    let attempt = 1
+
+    while (true) {
+      const controller = new AbortController()
+      const timeoutId = setTimeout(() => controller.abort(), 30000)
+
+      let res: Response
+      try {
+        res = await fetch(`${this.baseUrl}/scan`, {
+          method: 'POST',
+          headers,
+          body: JSON.stringify(body),
+          signal: controller.signal,
+        })
+      } catch (fetchErr) {
+        clearTimeout(timeoutId)
+        if (attempt < MAX_ATTEMPTS) {
+          const delay = Math.pow(2, attempt - 1) * 1000
+          attempt++
+          await new Promise((resolve) => setTimeout(resolve, delay))
+          continue
+        }
+        throw fetchErr
+      }
+
+      clearTimeout(timeoutId)
+
+      if (!res.ok) {
+        if (res.status === 429) {
+          const retryAfterHeader = res.headers.get('Retry-After')
+          const retryAfter = retryAfterHeader ? Math.ceil(parseFloat(retryAfterHeader)) : 60
+          throw new ApiError(429, 'Rate limited', retryAfter)
+        }
+
+        if (res.status >= 500 && res.status < 600 && attempt < MAX_ATTEMPTS) {
+          const delay = Math.pow(2, attempt - 1) * 1000
+          attempt++
+          await new Promise((resolve) => setTimeout(resolve, delay))
+          continue
+        }
+
+        const text = await res.text().catch(() => 'Unknown error')
+        throw new ApiError(res.status, text || `HTTP ${res.status}`)
+      }
+
+      const data = (await res.json()) as ScanResponse
+
+      const remaining = res.headers.get('X-RateLimit-Remaining')
+      const limit = res.headers.get('X-RateLimit-Limit')
+      const reset = res.headers.get('X-RateLimit-Reset')
+
+      const quota: ScanQuota | undefined =
+        remaining !== null && limit !== null && reset !== null
+          ? {
+              remaining: parseInt(remaining, 10),
+              limit: parseInt(limit, 10),
+              resetAt: parseInt(reset, 10) * 1000, // convert epoch seconds → ms
+            }
+          : undefined
+
+      return { ...data, quota }
+    }
+  }
+}
